For case notification, where you have "patient name" the cell asks for middle name and first name. In most countries, people use last name and first name. Actually, this is configurable and the user can change it on workspace configuration. But it is a good idea to set the "default" configuration of patient name to be last name and first name. I will change it.
Not sure if its logical for the previous BMU TB register date (entered oct 1 206) to be more recent than "date entered in SL treatment TB register" (entered apr. 13 2015). there should be a rule preventing this.
No it isn't, and it is a simple validation to implement. I will take care of it.

Under adverse events, the label "judged related to" appeared twice and its not intuitive about data required for entry
Sure, I will try to find a better name for this field.

Under previous TB treatment, the label "new previous TB treatment" is confusing. if its previous, it can't be new. Unless you mean the most recent previous treatment. Also drop down for months is in Portuguese.
Right, I will find a better name for the title and adjust the month lists.